Encapsulate CategoryList
